requirements.txt
nbtlib==1.12.1
pillow==10.3.0
numpy
